Move clickOutside to directives folder
