+++ src/directives/index.js
@@ -1,0 +1,2 @@
+export { default as clickOutside } from './click-outside';
+export { default as draggable } from './draggable';
