+++ src/util/index.js
@@ -1,4 +1,3 @@
-export { default as clickOutside } from './click-outside';
 export { default as countries } from './countries';
 export { default as getArrayFromLocale } from './get-array-from-locale';
 export { default as getCookie } from './get-cookie';
